+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Now supports the `detectable_format:constrained_response_with_argument` IFEval
+  constraint, being the same as `detectable_format:constrained_response` but with
+  a list of options to check for, rather than a hardcoded list of English options.
+
 ### Fixed
 
 - The `huggingface_hub` safetensors API has changed, so we did not fetch the number of